# Optimize forwarded calls