--- a/lib/rubocop/ast.rb
+++ b/lib/rubocop/ast.rb
@@ -5,6 +5,7 @@ require 'forwardable'
 require 'set'
 
 require_relative 'ast/ext/range'
+require_relative 'ast/utilities/simple_forwardable'
 require_relative 'ast/node_pattern/method_definer'
 require_relative 'ast/node_pattern'
 require_relative 'ast/node/mixin/descendence'