--- a/lib/rubocop/ast/node/mixin/collection_node.rb
+++ b/lib/rubocop/ast/node/mixin/collection_node.rb
@@ -4,7 +4,7 @@ module RuboCop
   module AST
     # A mixin that helps give collection nodes array polymorphism.
     module CollectionNode
-      extend Forwardable
+      extend SimpleForwardable
 
       ARRAY_METHODS =
         (Array.instance_methods - Object.instance_methods - [:to_a]).freeze