--- a/lib/rubocop/ast/node_pattern/compiler.rb
+++ b/lib/rubocop/ast/node_pattern/compiler.rb
@@ -9,7 +9,7 @@ module RuboCop
       # Doc on how this fits in the compiling process:
       #   /docs/modules/ROOT/pages/node_pattern.adoc
       class Compiler
-        extend Forwardable
+        extend SimpleForwardable
         attr_reader :captures, :named_parameters, :positional_parameters, :binding
 
         def initialize