Fixes [#312]

@Earlopain do you have a minute to review the PR? 🙏