@bbatsov 2024-09-02
Did you manage to get the same 3% speedup?