@marcandre 2024-09-03
Thanks!

Released as `1.32.2`

I'd recommend using that in `rubocop` too.